+++ src/grandorgue/GODocument.cpp
@@ -13,9 +13,9 @@
 #include "document-base/GOView.h"
 #include "gui/dialogs/GOMidiListDialog.h"
 #include "gui/dialogs/GOOrganSettingsDialog.h"
-#include "gui/dialogs/GOStopsDialog.h"
 #include "gui/dialogs/midi-event/GOMidiEventDialog.h"
 #include "gui/frames/GOFrame.h"
+#include "gui/frames/GOStopsWindow.h"
 #include "gui/panels/GOGUIPanel.h"
 #include "gui/panels/GOGUIPanelView.h"
 #include "gui/size/GOResizable.h"
@@ -191,15 +191,18 @@ void GODocument::ShowMidiList() {
 }
 
 void GODocument::ShowStops() {
-  if (!showWindow(GODocument::STOPS, NULL) && m_OrganController)
+  if (!showWindow(GODocument::STOPS, NULL) && m_OrganController) {
+    auto stopsWindow = new GOStopsWindow(
+      this,
+      nullptr,
+      m_OrganController->GetStopWindowSizeKeeper(),
+      *m_OrganController);
+
     registerWindow(
       GODocument::STOPS,
-      nullptr,
-      new GOStopsDialog(
-        this,
-        nullptr,
-        m_OrganController->GetConfig().m_DialogSizes,
-        *m_OrganController));
+      stopsWindow, // Otherwise GOStopsWindow::SyncState() wont be called
+      stopsWindow);
+  }
 }
 
 void GODocument::ShowMIDIEventDialog(