+++ src/grandorgue/gui/frames/GOStopsWindow.cpp
@@ -5,11 +5,15 @@
  * (https://www.gnu.org/licenses/old-licenses/gpl-2.0.html).
  */
 
-#include "GOStopsDialog.h"
+#include "GOStopsWindow.h"
 
 #include <wx/checkbox.h>
+#include <wx/display.h>
 #include <wx/intl.h>
+#include <wx/scrolwin.h>
 #include <wx/sizer.h>
+
+#include "gui/wxcontrols/go_gui_utils.h"
 
 #include "model/GOCoupler.h"
 #include "model/GODivisionalCoupler.h"
@@ -25,25 +29,23 @@ enum {
   ID_CHECKBOX = 200,
 };
 
-BEGIN_EVENT_TABLE(GOStopsDialog, GOSimpleDialog)
-EVT_CHECKBOX(ID_CHECKBOX, GOStopsDialog::OnElementChanging)
+BEGIN_EVENT_TABLE(GOStopsWindow, wxFrame)
+EVT_CHECKBOX(ID_CHECKBOX, GOStopsWindow::OnElementChanging)
+EVT_SHOW(GOStopsWindow::OnShow)
 END_EVENT_TABLE()
 
-GOStopsDialog::GOStopsDialog(
+GOStopsWindow::GOStopsWindow(
   GODocumentBase *doc,
   wxWindow *parent,
-  GODialogSizeSet &dialogSizes,
+  GOSizeKeeper &sizeKeeper,
   GOOrganModel &model)
-  : GOSimpleDialog(
-    parent,
-    wxT("Stops"),
-    _("Stops"),
-    dialogSizes,
-    wxEmptyString,
-    0,
-    wxCLOSE | wxHELP),
+  : wxFrame(parent, wxID_ANY, _("Stops")),
     GOView(doc, this),
+    r_SizeKeeper(sizeKeeper),
     r_model(model) {
+  SetIcon(get_go_icon());
+  wxScrolledWindow *sw = new wxScrolledWindow(this, wxID_ANY);
+
   wxBoxSizer *const mainSizer = new wxBoxSizer(wxHORIZONTAL);
   const unsigned nOdfManuals = model.GetODFManualCount();
   const unsigned globalSectionN = nOdfManuals;
@@ -53,7 +55,7 @@ GOStopsDialog::GOStopsDialog(
   for (unsigned i = model.GetFirstManualIndex(); i <= nOdfManuals; i++) {
     GOManual *const pManual = model.GetManual(i);
     wxStaticBoxSizer *const pSizer = new wxStaticBoxSizer(
-      wxVERTICAL, this, i < nOdfManuals ? pManual->GetName() : _("Globals"));
+      wxVERTICAL, sw, i < nOdfManuals ? pManual->GetName() : _("Globals"));
 
     sizers[i] = pSizer;
     mainSizer->Add(pSizer, 0, wxEXPAND | wxALL, 5);
@@ -109,7 +111,7 @@ GOStopsDialog::GOStopsDialog(
       if (
         sizerIndex >= 0 && sizerIndex <= (int)globalSectionN && pName
         && e.control->IsControlledByUser()) {
-        wxCheckBox *pCheckBox = new wxCheckBox(this, ID_CHECKBOX, *pName);
+        wxCheckBox *pCheckBox = new wxCheckBox(sw, ID_CHECKBOX, *pName);
 
         pCheckBox->SetClientData(e.control);
         mp_checkboxesByControl[e.control] = pCheckBox;
@@ -118,22 +120,42 @@ GOStopsDialog::GOStopsDialog(
       }
     }
   }
+  sw->SetScrollRate(5, 5);
+  sw->SetSizerAndFit(mainSizer);
 
-  LayoutWithInnerSizer(mainSizer);
+  // Set the optimal window size
+  SetClientSize(sw->GetVirtualSize());
+
+  // Lower the size if it does not fit the display.
+  wxRect max = wxDisplay(wxDisplay::GetFromWindow(this)).GetClientArea();
+  wxRect rect = GetRect();
+
+  // Check if the window fits the current display
+  if (!max.Contains(rect)) {
+    // Otherwise, check and correct width and height, The scrollbar will appear
+    if (rect.GetWidth() > max.GetWidth()) {
+      rect.SetWidth(max.GetWidth());
+      rect.SetX(max.GetX());
+    }
+    if (rect.GetHeight() > max.GetHeight()) {
+      rect.SetHeight(max.GetHeight());
+      rect.SetY(max.GetY());
+    }
+    SetSize(rect);
+  }
 }
 
-void GOStopsDialog::OnShow() {
-  GOSimpleDialog::OnShow();
-  r_model.RegisterControlChangedHandler(this);
+void GOStopsWindow::OnShow(wxShowEvent &event) {
+  if (event.IsShown()) {
+    r_SizeKeeper.ApplySizeInfo(*this);
+    r_model.RegisterControlChangedHandler(this);
+  } else {
+    r_model.UnRegisterControlChangedHandler(this);
+    r_SizeKeeper.CaptureSizeInfo(*this);
+  }
 }
 
-void GOStopsDialog::OnHide() {
-  // Don't move to the destructor becaus it may be called after ~GOOrganModel
-  r_model.UnRegisterControlChangedHandler(this);
-  GOSimpleDialog::OnHide();
-};
-
-void GOStopsDialog::OnElementChanging(wxCommandEvent &event) {
+void GOStopsWindow::OnElementChanging(wxCommandEvent &event) {
   wxCheckBox *pCheck = static_cast<wxCheckBox *>(event.GetEventObject());
   GOCombinationElement *pE
     = static_cast<GOCombinationElement *>(pCheck->GetClientData());
@@ -142,7 +164,7 @@ void GOStopsDialog::OnElementChanging(wxCommandEvent &event) {
     pE->SetCombinationState(event.IsChecked(), wxEmptyString);
 }
 
-void GOStopsDialog::ControlChanged(GOControl *pControl) {
+void GOStopsWindow::ControlChanged(GOControl *pControl) {
   auto pElement = dynamic_cast<GOCombinationElement *>(pControl);
 
   if (pElement) {
@@ -152,3 +174,5 @@ void GOStopsDialog::ControlChanged(GOControl *pControl) {
       it->second->SetValue(pElement->GetCombinationState());
   }
 }
+
+void GOStopsWindow::SyncState() { r_SizeKeeper.CaptureSizeInfo(*this); }